refactor enum Asserton
