+++ primitives/core/src/assertion.rs
@@ -16,14 +16,13 @@
 
 // This file includes the predefined rulesets and the corresponding parameters
 // when requesting VCs.
-//
-// See: https://www.notion.so/litentry/Expected-parameters-in-predefined-rulesets-14f74928aa2b43509167da12a3e75507
 
+use crate::Balance;
 use codec::{Decode, Encode, MaxEncodedLen};
-use core_primitives::Balance;
 use scale_info::TypeInfo;
 use sp_runtime::{traits::ConstU32, BoundedVec};
 
+// pub type Balance = u128;
 type MaxStringLength = ConstU32<64>;
 pub type ParameterString = BoundedVec<u8, MaxStringLength>;
 pub type Network = BoundedVec<u8, MaxStringLength>;
@@ -35,10 +34,7 @@ pub enum Assertion {
 	A2(ParameterString),                                   // (guild_id)
 	A3(ParameterString, ParameterString, ParameterString), // (guild_id, channel_id, role_id)
 	A4(Balance),                                           // (minimum_amount)
-	A5(ParameterString, ParameterString),                  /* (twitter_account, tweet_id), -
-	                                                        * TODO: unsupported,
-	                                                        * until Twitter new API policy is
-	                                                        * published */
+	A5(ParameterString, ParameterString),                  // (twitter_account, tweet_id)
 	A6,
 	A7(Balance),           // (minimum_amount)
 	A8(AssertionNetworks), // litentry, litmus, polkadot, kusama, khala, ethereum
@@ -47,3 +43,16 @@ pub enum Assertion {
 	A11(Balance), // (minimum_amount)
 	A13(u32),     // (Karma_amount) - TODO: unsupported
 }
+
+pub const ASSERTION_NETWORKS: [&str; 6] =
+	["litentry", "litmus", "polkadot", "kusama", "khala", "ethereum"];
+
+pub const ASSERTION_FROM_DATE: [&str; 7] = [
+	"2017-01-01",
+	"2018-01-01",
+	"2019-01-01",
+	"2020-01-01",
+	"2021-01-01",
+	"2022-01-01",
+	"2023-01-01",
+];
